+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- add `--cmeel-env` option to `docker` subcommand to forward `CMEEL_*` environment variables
+
 ## [v0.42.1] - 2023-05-03
 
 - `[project]` section of `pyproject.toml`: